raspi/requirements.txt
# AdaptLight Python Requirements
# Install with: pip install -r requirements.txt

# Hardware control
RPi.GPIO==0.7.1          # GPIO control (Raspberry Pi only)
lgpio                    # Modern GPIO library for Raspberry Pi 5
rpi-lgpio                # lgpio wrapper for compatibility
adafruit-circuitpython-neopixel-spi  # NeoPixel SPI control (Pi 5 compatible)
adafruit-blinka          # CircuitPython compatibility layer
gpiozero==2.0.1          # High-level GPIO interface

# AWS
boto3==1.34.0            # AWS SDK for Python
botocore==1.34.0

# Configuration
PyYAML==6.0.1            # YAML config files

# OpenAI
openai>=1.52.0           # OpenAI API client

# Audio (for voice input)
pyaudio==0.2.14          # Audio capture
sounddevice==0.4.6       # Alternative audio interface
vosk==0.3.45             # Offline speech recognition (optional)
numpy>=1.26.2            # Numerical operations for audio processing

# Scheduling
APScheduler==3.10.4      # Task scheduling for uploads

# Utilities
requests==2.31.0         # HTTP requests
